+++ example/Nav.js
@@ -1,6 +1,5 @@
 import React, {Component} from 'react'
-import {PropTypes} from 'prop-types'
-import {Link} from 'react-router'
+import {Link} from 'react-router-dom'
 
 class Nav extends Component {
     constructor(props) {