+++ example/Summary.js
@@ -47,6 +47,7 @@ class Summary extends Component{
                                moveEvent={this.moveEvent}
                                newEvent={this.newEvent}
                                leftCustomHeader={leftCustomHeader}
+                               toggleExpandFunc={this.toggleExpandFunc}
                     />
                 </div>
             </div>
@@ -162,6 +163,13 @@ class Summary extends Component{
         this.setState({
             viewModel: schedulerData
         })
+    }
+
+    toggleExpandFunc = (schedulerData, slotId) => {
+        schedulerData.toggleExpandStatus(slotId);
+        this.setState({
+            viewModel: schedulerData
+        });
     }
 }